Refactor project structure to use packages
This moves MainActivity into its own "main" package under the new "screens" package.
