+++ app/src/main/res/layout/activity_main.xml
@@ -9,7 +9,7 @@
     android:paddingTop="@dimen/activity_vertical_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingBottom="@dimen/activity_vertical_margin"
-    tools:context=".MainActivity">
+    tools:context=".screens.main.MainActivity">
 
     <com.google.android.material.textfield.TextInputLayout
         android:id="@+id/phone_number_input_layout"
