+++ app/src/main/res/menu/menu.xml
@@ -1,7 +1,7 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
       xmlns:app="http://schemas.android.com/apk/res-auto"
       xmlns:tools="http://schemas.android.com/tools"
-      tools:context=".MainActivity">
+      tools:context=".screens.main.MainActivity">
     <item
         android:id="@+id/action_about"
         android:title="@string/menu_item_about"
